Fix: Changing values in ArrayField is impossible
The way this was implemented lead to a copy of the displayed values
being made when the UI was generated. This collection would always
override the modified values when the ArrayField was refreshed. To fix
this we can instead create the collection in the closure, to make sure
up-to-date values get displayed and the bound values do not get
overridden with an old state.